requirements.txt
openai==1.51.2
Pillow==10.2.0
transformers==4.45.2
opencv-python==4.6.0.66
ipykernel==6.15.2
numpy==1.25.0
pandas
httpx==0.27.2
accelerate>=0.26.0
timm==1.0.10
rich==13.3.2